requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
pytesseract
Pillow
opencv-python-headless
numpy
scikit-image
spacy
pyspellchecker
language-tool-python
nltk
python-dotenv
aiofiles
loguru
pytest
